+++ docs/about/release-notes.rst
@@ -28,6 +28,15 @@ Release Notes
    Neil Vaytet :sup:`a`\ ,
    and Jan-Lukas Wynen :sup:`a`
 
+v0.3.3 (October 2022)
+---------------------
+
+Bugfixes
+~~~~~~~~
+
+* Fix :meth:`scippnexus.Field.dims` and :meth:`scippnexus.Field.shape` to consistently return tuples instead of lists `#62 <https://github.com/scipp/scippnexus/pull/62>`_.
+* Fix :meth:`scippnexus.NXobject.__getitem__` to return children with correctly set up field dimensions when indexed with a class name `#62 <https://github.com/scipp/scippnexus/pull/62>`_.
+
 v0.3.0 (September 2022)
 -----------------------